feat(plugins): Intro. `didResolveSource` to indicate availability of "source".
This PR targets https://github.com/apollographql/apollo-server/pull/3988 but
was prompted by a need identified during review in the following comment:

https://github.com/apollographql/apollo-server/pull/3998/files#r414911049

As noted, this is tangible clunkiness which could make the implementation of
the `ensurePreflight` invocations unnecessary.  That defensiveness was only
necessary because the extensions API - which #3988 + #3998 aims to render
unnecessary - guaranteed the presence of the "source" (the text of the
incoming operation) prior to invoking its `requestDidStart` method (which
was actually a bit after the request had started, in reality).

Hopefully this proves to be a useful life-cycle for other cases!