Add comments about preservation of legacy behaviors in request pipeline.
These comments hope to better explain the behavior which was present in the
2.x series of Apollo Server prior to the introduction of
`didEncounterErrors`.

The behavior which this commentary applies to -- in which we `throw` prior
to finishing subsequent errors, has been this way since AS2 was released and
was a pattern which we locked ourselves into for the duration of the AS2
release when we put out the request pipeline plugins.

It should _not_ stay this way in the future, and in AS3 we will resolve this.