Remove integrated subscriptions in favor of external dependency.
The idea behind removing subscriptions here is not to dismiss their
importance.  We will be providing documentation in a migration guide on how
to achieve parity of the implementation we provided in Apollo Sever 2.0.

Prior to this commit, by default, subscriptions are on for all users of
Apollo Server.  That includes users who may not actually need "subscription"
behavior, users that are not using `Subscription` types in their schema, and
users that we haven't yet provided enough information on the many
considerations that must be made in order to implement subscriptions
successfully.

Apollo Server 2.0 brought a work-in-progress subscriptions model into Apollo
Server.  That model included the `PubSub` model provided by
`graphql-subscriptions` and the WebSocket transport provided by
`subscriptions-transport-ws`.

While we thought we'd be able to build on this model, we have had to realize
that it requires a lot more effort to get it right.  We believe we can put
together a much better story for subscriptions, but we are unable to do it
within the current model we brought in Apollo Server 2.0.

Further, by having subscriptions on by default, we have led many users down
a path of thinking that the subscriptions implementation we have was stable,
scalable, and being heavily invested in.  In fact, most users don't even
need subscriptions and actually benefit from a story using entirely
different primitives.  To this day, subscriptions in Apollo Server do not
use the same request pipeline as the central request pipeline that we've
built in Apollo Server 2.0  That means they don't support plugins, they
don't support metrics reporting, they aren't supported within federation,
and they don't work with the operation registry.  They don't even work with
the legacy `graphql-extensions` API.

We are going to put time into re-thinking subscriptions, but the first step
is removing the default implementation that we offered which can be
reimplemented with a few lines of code for users that absolutely necessitate
it.  Those users have stories and perspectives that we need to consider
based on their demonstrated use cases for subscriptions, and we need those
stories to guide a future, better implementation.

A separate commit will bring the documentation for a migration story.