Revert "Release"
This reverts commit 922af93026976bf58df10b8ee9e5f0461f6f5cd3.

As noted in the commit prior to this one: df32baf, TS errors were
introduced by cherry-picking 85c18db (which was targeting a
different branch previously, an easy mistake to make).

Associated git tags were removed from the existing release commit
with the intention that they'll be used in the release commit
following this one.