tests(op-reg): Use the `reportTiming` option to disable reporting.
Previously, we were leveraging `sendReportsImmediately` to ... do exactly
that even though these tests don't need to send reports _at all_!.  This was
purely a hack to make sure that we don't need to wait for the Engine
Reporting Agent to eventually fail (after 5 seconds or whatever its
reporting interval is).

This new `reportTiming` option is available with the below-linked PR, so we
can now clean this up a bit.  This has the added advantage of not hitting
our actual reporting endpoint with an invalid API key, and having the logs
spew errors about failures in delivering those doomed reports.

Ref: https://github.com/apollographql/apollo-server/pull/3918