Enable tests for Node.js 14 on CircleCI.
The Node.js 14 line will enter LTS (and be assigned its codename!) in
October 2020.  We are already testing Node.js 14 on the v3 branch, but we
should start some early tests on Node.js 14 for the v2 series and also
within introducing this the Gateway is only tested on Node.js 12 right now
(intentionally).