feat(gateway): New federation composition format (#4405)
The output of composition as it stands is not a portable one. In its current
form, the output of composition is an executable GraphQLSchema object whose
types and fields are annotated via the extensions field.

This is problematic for the future of composition at Apollo - as we move query
planning to Rust, moving away from this JS native object and towards a portable
format (SDL) is ideal, and also allows us to:

* parse and consume this format more easily with tooling
* create / consume this format in any language which can parse valid GraphQL
* standardize the format for other implementors

All of that being said, this commit creates an additional output to
composition - the printed SDL of the new composition format. We're calling this
CSDL C(omposed)SDL to distinguish that this SDL ought to include all of the
annotations / metadata necessary about the composed graph in order to perform
query planning and execution.