fix(federation): Create new @key validations (#4498)
A few cases of invalid @key directive arrangements are currently possible
with our existing validation suite. This PR adds 3 new validations
to ensure that users are informed of these invalid states and block
composition appropriately.

* Owning type must specify a @key
* Extending types can't specify multiple @keys
* Extending types must use a @key specified by the owning type

Additionally:
These invalid arrangements can result in runtime errors when printing CSDL,
nor is it really valid to return an "attempted" print of CSDL, so the result
of `composedSdl` is now `undefined` in the case that there are composition
errors.