Switch to a fork of graphql-upload
This fork is identical to graphql-upload@8.1.0 except that it allows graphql@15
as a peer dependency (and bundles the TypeScript type definitions directly).

For #4865.

Note that we do plan to remove this integration in Apollo Server 3, but for now
we want you to be able to combine AS2 with graphql@15 and npm@7.