limit to engines.npm 6.x
This is now used to control the npm version that is used for installing
updates.  I fully support us making the migration to npm@7 at this point,
but it should be done in an structured, intentional manner and this just
makes the status-quo more clear to tooling that is just going to use the
latest version if it doesn't know better.