CI: new job which fails if a file contains FIXME (#5121)
When working on a big PR, sometimes you have something you know you need to fix
before merging. With this CI job, just add a FIXME comment and you'll be forced
to address it before all CI can pass.

We intentionally allow CI to make tarballs for a commit and publish npm packages
even with FIXME, because you may want to test out a build of an incomplete PR.

Change existing TODO comments to FIXME. Document this requirement in
CONTRIBUTING.md and make the text about PRs with failing tests a bit gentler.