Remove 'tracing' constructor option and apollo-server-tracing (#5193)
This option/package enabled a format for execution traces that was
primarily used for engineproxy (replaced 3 years ago) as well as some
versions of Playground. This `tracing` response extension was very
verbose as it used a flat list for resolvers rather than a nested one
like the `ftv1` extension used by ApolloServerPluginInlineTrace. It is
essentially an attractive nuisance: an innocuously named option that
increases the size of your responses without powering any actively
maintained tools.

We have not yet made any backwards-incompatible changes in the AS plugin
API, so you still can use the AS2 `apollo-server-tracing` plugin
directly with AS3 servers; we just won't publish the package any more
and we won't provide a boolean that installs it for you.

Part of #5158.