cache control: update requestContext.overallCachePolicy continuously
I got stressed out about whether it mattered that the cache control
plugin used to be almost always first in the plugins list. But why does
this matter? It matters because the plugin only updated
overallCachePolicy at a few specific moments in the request pipeline.

This commit simplifies the model by updating
requestPolicy.overallCachePolicy each time a field finishes resolving.

In order to keep the existing unit tests of "addHint" working, I moved
"addHint" to a PolicyUpdater object. Note that there's one extra boolean
of state to track: there's a difference between
"`requestContext.overallCachePolicy` is undefined because we've just
started" (in which case the first hint should override the lack of a
policy) and "`requestContext.overallCachePolicy` is undefined because
we've actually seen something that says the response can't be cached",
in which case we should never change overallCachePolicy. That state gets
tracked inside PolicyUpdater.

I respected the existing logic that never touches overallCachePolicy if
it is set by something outside of the cache control plugin. That said,
the only use case I know of for that is the response cache's
responseForObject, and when it sets overallCachePolicy the query doesn't
get executed and the cache control plugin never tries to write to
overallCachePolicy anyway. So it's not clear how important that logic
is.

Note that we're now careful to call `addHint` only once per field (when
it's done resolving), even with `setCacheHint`. That's because if you
call `info.cacheControl.setCacheHint({maxAge: 10})` on a field with
`@cacheControl(maxAge: 5)`, the 10 does override the 5 and so we can't
have already incorporated the 5 into overallCachePolicy.